do not require key for Optional in Struct
Optional is equal to Select(subcon, Pass), so there is no practical
difference between sending None to subcon explicitly and having it
fail, and sending None implicitly when no value is provided.
So technically speaking, flagbuildnone is true for this particular form
of Select even in cases when it is false for subcon.

This change then allows using Optional as if it were
Default(Optional, None)