Fix catch-all exception handling for debugging
When debugging issue #888, I found that this catch-all exception handler
was hiding a real bug in construct when replacing it with its own
exception.

A nice pattern to avoid this hiding is the `raise ... from ...` which
will cause Python to carry through the original exception information,
which then appears as part of the full traceback when it is printed.

I suggest that this pattern be used for all cases of catch-all exception
"wrapping", but for now I do this specifically just to assist with the
issue being addressed.